refactor: move the button component to buttons module for clarity (#18)
