+++ tests/unit/nicemldashboard/basecomponents/test_buttons.py
@@ -1,12 +1,14 @@
 import pytest
 from unittest.mock import MagicMock
 
-from nicemldashboard.basecomponents.buttons import (
-    SidebarToggleButton,
+from nicemldashboard.basecomponents.buttons.sidebartogglebutton import (
     ARROW_RIGHT_ICON,
     ARROW_LEFT_ICON,
     MAX_WIDTH,
     MIN_WIDTH,
+)
+from nicemldashboard.basecomponents.buttons.sidebartogglebutton import (
+    SidebarToggleButton,
 )
